설계한대로 구조 변경
공격아이템,능력치변경아이템,날개아이템으로 분화
PlayerInfo에서 각 아이템 습득시 동작 구현
능력치 변경아이템 상속으로 구현
GameManager에서 모든 아이템 정보 갖고있을것.

+++ Assets/02.Scripts/Enemies/EnemyInfo.cs
@@ -6,7 +6,7 @@ using UnityEngine.AI;
 public class EnemyInfo : MonoBehaviour
 {
     const string PLAYER = "PLAYER";
-    public int healthPoint;
+    public float healthPoint;
     public int enemyDamage;
     private bool isDead;
     public bool IsDead
